rewrite: Support refresh and serve cached by default from all, article and the checking endpoints.
