+++ src/views/statistics/write_view.py
@@ -1,0 +1,13 @@
+from src.views.statistics import StatisticsView
+
+
+class StatisticsWriteView(StatisticsView):
+    def __setup_io__(self):
+        raise NotImplementedError()
+
+    def __handle_valid_job__(self):
+        raise NotImplementedError()
+
+    def __read_from_cache__(self):
+        self.__setup_io__()
+        self.io.read_from_disk()
